+++ CHANGELOG.md
@@ -4,6 +4,10 @@ https://github.com/marco-m/otium
 
 ## UNRELEASED v0.1.7
 
+### New
+
+- When invoked with --doc-only, an otium procedure prints the documentation instead of running.
+
 ### Breaking
 
 - Procedure.Execute: change signature to `Execute(args []string) error`, to ease testing. Normally you would use this as `pcd.Execute(os.Args)`.